+++ crates/compiler/src/errors.rs
@@ -11,9 +11,9 @@ use crate::bytecode::{
     InstrumentId, LoopCount, NoiseFrequency, Pan, PanSlideAmount, PanSlideTicks,
     PanbrelloAmplitude, PanbrelloQuarterWavelengthInTicks, PlayPitchPitch, PortamentoSlideTicks,
     PortamentoVelocity, RelativeEchoFeedback, RelativeEchoVolume, RelativeFirCoefficient,
-    RelativePan, RelativeVolume, Transpose, TremoloAmplitude, TremoloQuarterWavelengthInTicks,
-    VibratoDelayTicks, VibratoPitchOffsetPerTick, VibratoQuarterWavelengthInTicks, Volume,
-    VolumeSlideAmount, VolumeSlideTicks,
+    RelativeMainVolume, RelativePan, RelativeVolume, Transpose, TremoloAmplitude,
+    TremoloQuarterWavelengthInTicks, VibratoDelayTicks, VibratoPitchOffsetPerTick,
+    VibratoQuarterWavelengthInTicks, Volume, VolumeSlideAmount, VolumeSlideTicks,
 };
 use crate::command_compiler::channel_bc_generator::MAX_NO_LOOP_TICKS;
 use crate::command_compiler::commands::{
@@ -187,6 +187,7 @@ pub enum ValueError {
     NoRelativeTransposeSign,
     NoDetuneValueSign,
     NoDetuneCentsSign,
+    NoRelativeMainVolumeSign,
     NoRelativeEchoVolumeSign,
     NoRelativeEchoFeedbackSign,
     NoRelativeFirCoefficientSign,
@@ -222,6 +223,8 @@ pub enum ValueError {
     MainVolumeOutOfRange(i32),
     MainVolumeOutOfRangeU32(u32),
     MainVolumeHexOutOfRange(u32),
+    RelativeMainVolumeOutOfRange(i32),
+    RelativeMainVolumeOutOfRangeU32(u32),
 
     EchoEdlLargerThanMaxEdl { edl: EchoEdl, max_edl: EchoEdl },
     EchoEdlOutOfRange(u32),
@@ -298,6 +301,7 @@ pub enum ValueError {
     NoRelativeEchoVolume,
     NoEchoFeedback,
     NoMainVolume,
+    NoRelativeMainVolume,
     NoFirTap,
     NoRelativeEchoFeedback,
     NoFirCoefficient,
@@ -1151,6 +1155,9 @@ impl Display for ValueError {
             Self::NoDetuneCentsSign => {
                 write!(f, "missing + or - in detune cents")
             }
+            Self::NoRelativeMainVolumeSign => {
+                write!(f, "missing + or - in relative main volume")
+            }
             Self::NoRelativeEchoVolumeSign => {
                 write!(f, "missing + or - in relative echo volume")
             }
@@ -1206,6 +1213,12 @@ impl Display for ValueError {
             Self::MainVolumeOutOfRangeU32(v) => out_of_range!("main volume", v, MainVolume),
             Self::MainVolumeHexOutOfRange(v) => {
                 write!(f, "cannot parse main volume: ${v:x} is not a byte value")
+            }
+            Self::RelativeMainVolumeOutOfRange(v) => {
+                out_of_range!("relative main volume", v, RelativeMainVolume)
+            }
+            Self::RelativeMainVolumeOutOfRangeU32(v) => {
+                out_of_range!("relative main volume", v, RelativeMainVolume)
             }
 
             Self::EchoEdlLargerThanMaxEdl { edl, max_edl } => {
@@ -1346,6 +1359,7 @@ impl Display for ValueError {
             Self::NoRelativeEchoVolume => write!(f, "no relative echo volume"),
             Self::NoEchoFeedback => write!(f, "no echo feedback value"),
             Self::NoMainVolume => write!(f, "no main volume value"),
+            Self::NoRelativeMainVolume => write!(f, "no relative main volume"),
             Self::NoFirTap => write!(f, "no fir tap"),
             Self::NoRelativeEchoFeedback => write!(f, "no relative echo feedback"),
             Self::NoFirCoefficient => write!(f, "no fir coefficient"),